+++ clang/lib/AST/ASTImporter.cpp
@@ -10613,6 +10613,7 @@ ASTNodeImporter::ImportAPValue(const APValue &FromValue) {
   switch (FromValue.getKind()) {
   case APValue::None:
   case APValue::Indeterminate:
+  case APValue::Erroneous:
   case APValue::Int:
   case APValue::Float:
   case APValue::FixedPoint: