+++ clang/lib/AST/ItaniumMangle.cpp
@@ -6430,6 +6430,7 @@ static bool isZeroInitialized(QualType T, const APValue &V) {
   switch (V.getKind()) {
   case APValue::None:
   case APValue::Indeterminate:
+  case APValue::Erroneous:
   case APValue::AddrLabelDiff:
     return false;
 
@@ -6583,6 +6584,7 @@ void CXXNameMangler::mangleValueInTemplateArg(QualType T, const APValue &V,
   switch (V.getKind()) {
   case APValue::None:
   case APValue::Indeterminate:
+  case APValue::Erroneous:
     Out << 'L';
     mangleType(T);
     Out << 'E';