+++ clang/lib/AST/MicrosoftMangle.cpp
@@ -1940,6 +1940,7 @@ void MicrosoftCXXNameMangler::mangleTemplateArgValue(QualType T,
   switch (V.getKind()) {
   case APValue::None:
   case APValue::Indeterminate:
+  case APValue::Erroneous:
     // FIXME: MSVC doesn't allow this, so we can't be sure how it should be
     // mangled.
     if (WithScalarType)