+++ clang/lib/Sema/AnalysisBasedWarnings.cpp
@@ -959,8 +959,34 @@ static void DiagUninitUse(Sema &S, const VarDecl *VD, const UninitUse &Use,
                           bool IsCapturedByBlock) {
   bool Diagnosed = false;
 
+  // [basic.indet]/p1.1:
+  //  - If the object has dynamic storage duration, or is the object associated
+  //    with a variable or function parameter whose first declaration is marked
+  //    with the [[indeterminate]] attribute ([dcl.attr.indet]), the bytes have
+  //    indeterminate values;
+  //
+  //  - otherwise, the bytes have erroneous values, where each value is
+  //  determined
+  //    by the implementation independently of the state of the program.
+  //
+  // If variable has automatic storage duration and does
+  // not have [[indeterminate]], reading it is erroneous behavior (not
+  // undefined). However, we still warn about it.
+  bool IsErroneousBehavior = S.getLangOpts().CPlusPlus26 &&
+                             VD->hasLocalStorage() &&
+                             !VD->hasAttr<IndeterminateAttr>();
   switch (Use.getKind()) {
   case UninitUse::Always:
+    if (IsErroneousBehavior &&
+        !S.Diags.isIgnored(diag::warn_erroneous_behavior_uninitialized_read,
+                           Use.getUser()->getBeginLoc())) {
+      S.Diag(Use.getUser()->getBeginLoc(),
+             diag::warn_erroneous_behavior_uninitialized_read)
+          << VD->getDeclName() << Use.getUser()->getSourceRange();
+      S.Diag(VD->getLocation(), diag::note_variable_erroneous_init)
+          << VD->getDeclName();
+      return;
+    }
     S.Diag(Use.getUser()->getBeginLoc(), diag::warn_uninit_var)
         << VD->getDeclName() << IsCapturedByBlock
         << Use.getUser()->getSourceRange();
@@ -3255,7 +3281,9 @@ void clang::sema::AnalysisBasedWarnings::IssueWarnings(
       !Diags.isIgnored(diag::warn_sometimes_uninit_var, D->getBeginLoc()) ||
       !Diags.isIgnored(diag::warn_maybe_uninit_var, D->getBeginLoc()) ||
       !Diags.isIgnored(diag::warn_uninit_const_reference, D->getBeginLoc()) ||
-      !Diags.isIgnored(diag::warn_uninit_const_pointer, D->getBeginLoc())) {
+      !Diags.isIgnored(diag::warn_uninit_const_pointer, D->getBeginLoc()) ||
+      !Diags.isIgnored(diag::warn_erroneous_behavior_uninitialized_read,
+                       D->getBeginLoc())) {
     if (CFG *cfg = AC.getCFG()) {
       UninitValsDiagReporter reporter(S);
       UninitVariablesAnalysisStats stats;