+++ clang/lib/Sema/SemaTemplate.cpp
@@ -8134,6 +8134,7 @@ static Expr *BuildExpressionFromNonTypeTemplateArgumentValue(
 
   case APValue::None:
   case APValue::Indeterminate:
+  case APValue::Erroneous:
     llvm_unreachable("Unexpected APValue kind.");
   case APValue::LValue:
   case APValue::MemberPointer: