+++ clang/docs/ReleaseNotes.rst
@@ -107,6 +107,7 @@ C++ Language Changes
 
 C++2c Feature Support
 ^^^^^^^^^^^^^^^^^^^^^
+- Clang now supports `P2795R5 <https://wg21.link/p2795r5>`_ Erroneous behaviour for uninitialized reads.
 
 C++23 Feature Support
 ^^^^^^^^^^^^^^^^^^^^^